refactor: split app/ui and add architecture docs
