+++ internal/app/errors.go
@@ -1,0 +1,5 @@
+package app
+
+import "errors"
+
+var errTmuxUnavailable = errors.New("tmux service unavailable")
